Module peregrine::libc::memory

Structs

IOVec
Defines an I/O Vector that contains a pointer and size.

Functions

__aeabi_memclr

__aeabi_memset

memcmp
Compares two regions of memory.

memcmp_t
Compare regions of memory with a generic iterator size.

memcpy
Copies one region of memory to another.

memset
Set a region of memory to a given byte value.