Struct peregrine::libc::memory::IOVec

pub struct IOVec {
    pub ptr: *const u8,
    pub size: usize,
}

Defines an I/O Vector that contains a pointer and size.

Fields
ptr: *const u8 size: usize
Methods

impl IOVec

fn new(ptr: *const u8, size: usize) -> IOVec

Instantiates a new IOVec given the pointer and size.

fn from_addr(addr: u32, size: usize) -> IOVec

Instantiates a new IOVec given a u32 address and size

fn from_range<T>(begin: *const T, end: *const T) -> IOVec

Instantiates a new IOVec given two pointers

fn convert_ptr<T>(&self) -> *const T

Casts the contained pointer to a pointer of the given generic type.

fn convert_mut_ptr<T>(&self) -> *mut T

Casts the contained pointer to a pointer of the given generic type as well as making it mutable.

fn as_mut(&self) -> *mut u8

Returns the pointer as a mutable pointer.