Struct peregrine::libc::structures::locks::SpinLock

pub struct SpinLock { /* fields omitted */ }

Methods

impl SpinLock

fn new() -> SpinLock

fn is_locked(&self) -> bool

fn acquire(&mut self)

Acquires the lock. If the lock is currently held, wait for it to clean and then take it.

fn release(&mut self) -> Result<(), &'static str>

Release the lock, but return error if not locked.

fn wait(&self)

Waits for the lock to be released, but does not acquire it.