Struct peregrine::libc::structures::Bitmap

pub struct Bitmap {
    // some fields omitted
}

Methods

impl Bitmap

fn new(addr: *const u8, size: usize) -> Bitmap

Create a new IOVec from the info, and initialize a bitmap

fn from_iov(iov: IOVec) -> Bitmap

Consume the IOVec and return the initialized bitmap

fn count(&self) -> usize

Get the total number of bits in the bitmap.

fn used(&self) -> usize

Get the number of set bits in the bitmap.

fn free(&self) -> usize

Get the number of unset bits in the bitmap.

fn is_set(&self, index: usize) -> bool

Get the state of the Nth bit.

fn set(&mut self, index: usize) -> Result<(), &'static str>

Sets the bit at the given index.

An error is returned if the index is out of the bitmap's bounds.

fn checked_set(&mut self, index: usize) -> Result<(), &'static str>

Sets the bit at the given index.

An error is returned if the index is out of the bitmap's bounds or if the bit was already set.

fn set_range(&mut self, index: usize, count: usize) -> Result<(), &'static str>

Sets the bits starting at the given index through the given range.

An error is returned if the index is out of the bitmap's bounds.

fn clear(&mut self, index: usize) -> Result<(), &'static str>

Clears the bit at the given index.

An error is returned if the index is out of the bitmap's bounds.

fn checked_clear(&mut self, index: usize) -> Result<(), &'static str>

Clears the bit at the given index.

An error is returned if the index is out of the bitmap's bounds or if the bit was already unset.

fn find(&self, count: usize) -> Option<usize>

Search the bitmap for count contiguous unset bits. If found, return the block number that starts the region.

fn find_and_set(&mut self, count: usize) -> Result<usize, &'static str>

Search the bitmap for count contiguous unset bits. If found, set them.