Struct peregrine::libc::structures::RingBuffer

pub struct RingBuffer<T: Sized> { /* fields omitted */ }

Methods

impl<T> RingBuffer<T>

fn new(mem: IOVec) -> RingBuffer<T>

Creates a new ring buffer spanning the region defined by the IOVec.

NOTE: because pointers are indexed using signed integers, you should not create a buffer larger than 231 elements.

fn from(mem: &[T]) -> RingBuffer<T>

Creates a ring buffer from the given slice using the slice's .len() function and a pointer to the first index.

NOTE: because pointers are indexed using signed integers, you should not create a buffer larger than 231 elements.

fn size(&self) -> usize

Returns the (1-indexed) number of items the ring buffer can hold.

fn get(&self, index: usize) -> Result<&T, &'static str>

Get the value at the "relative index" of the ring buffer.

By "relative index", the API means the following contract:

buff.get(0) will give the oldest value in the buffer

buff.get(buff.size()-1) will give the newest element in the list.

NOTE: offsetting using isize effectively limits the actual max size of the ring buffer to 231.

Errors are generated on:

Using an index either:

Beyond the size of the buffer

Beyond the number of inserted items (... only if the buffer has not wrapped yet)

The calculated pointer (then converted to ref) to the index is nullptr. This can happen when:

The user creates a ring buffer starting at 0x0usize and calls .get(0)

The buffer is located high in memory such that the offsetting wraps the number space to 0.

Examples

use peregrine::libc::structures::RingBuffer;

// create a "memory region" to use as the buffer
let arr = [0u32; 16];
let mut buff = RingBuffer::from(&arr);

// fill the buffer twice
for i in 0..(arr.len()*2) { buff.push(i as u32); }

// assert our values
assert_eq!(
    arr.len() as u32,
    *buff.get(0).expect("failed to get index 0"),
    "incorrect value for .get(0) case"
);
assert_eq!(
    ((arr.len()*2)-1) as u32,
    *buff.get(buff.size()-1).expect("failed to get last index"),
    "incorrect value for .get(last) case"
);

// assert they are equal to the `newest()` and `oldest()` helper functions
assert_eq!(
    *buff.newest().expect("failed to get newest"),
    *buff.get(buff.size()-1).expect("failed to get last index"),
    "incorrect value for get-newest case"
);
assert_eq!(
    *buff.oldest().expect("failed to get oldest"),
    *buff.get(0).expect("failed to get index 0"),
    "incorrect value for get-oldest case"
);Run

fn push(&mut self, val: T)

Pushes a value into the ring buffer, overwriting the oldest value in the buffer.

If the boundry of the buffer is hit, the internal iterator wraps the buffer, restarting at "index 0".

Examples

use peregrine::libc::structures::RingBuffer;

// create a "memory region" to use as the buffer
let arr = [0u32; 16];
let mut buff = RingBuffer::from(&arr);

// fill the buffer exactly
for i in 0..(arr.len() as u32) { buff.push(i); }

// expect nothing to be overwritten yet
assert_eq!(0, *buff.get(0).unwrap());
assert_eq!((arr.len() as u32) - 1, *buff.get(buff.size()-1).unwrap());

// now loop, and overwrite
buff.push(0xDEADBEEF);
assert_eq!(0xDEADBEEF, *buff.newest().unwrap());
assert_eq!(0xDEADBEEF, arr[0]);Run

fn newest(&self) -> Result<&T, &'static str>

Returns the newest element in the buffer.

NOTE: calling .newest() on an empty buffer will result in an error.

Examples

use peregrine::libc::structures::RingBuffer;

// create a "memory region" to use as the buffer
let arr = [0u32; 16];
let mut buff = RingBuffer::from(&arr);

// fill and assert
for i in 0..(arr.len() as u32) {
    if i > 0 { assert_eq!(i-1, *buff.newest().unwrap()); }
    buff.push(i);
    assert_eq!(i, *buff.newest().unwrap());
}Run

fn oldest(&self) -> Result<&T, &'static str>

Returns the oldest element in the buffer.

NOTE: calling .oldest() on an empty buffer will result in an error.

Examples

use peregrine::libc::structures::RingBuffer;

// create a "memory region" to use as the buffer
let arr = [0u32; 16];
let mut buff = RingBuffer::from(&arr);

// fill and assert
for i in 0..(arr.len() as u32) {
    buff.push(i);
    assert_eq!(0, *buff.oldest().unwrap());
}

// roll the buffer and do a new assert
buff.push(0xDEADBEEF);
assert_eq!(1, *buff.oldest().unwrap());Run