Struct peregrine::mcus::cortexm4::core::fpu::Unit

pub struct Unit(pub *const u8);

Pointer wrapper that implements the Unit operations

Methods

impl Unit

fn read_context(&self) -> u32

Reads the contents (as u32) of the context register at relative address 0x0

fn disable_automatic_state_saving(&self)

Writes the value 0x1 to the context register at relative address 0x0 and offset 31 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_automatic_state_saving(&self)

Writes the value 0x1 to the context register at relative address 0x0 and offset 31 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_lazy_state_saving(&self)

Writes the value 0x1 to the context register at relative address 0x0 and offset 30 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_lazy_state_saving(&self)

Writes the value 0x1 to the context register at relative address 0x0 and offset 30 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn read_context_address(&self) -> u32

Reads the contents (as u32) of the context_address register at relative address 0x4

fn needs_init(&self) -> bool

states the Unit will not be initialized

fn init(&self)

Should not be called as it does nothing. Here to support the init interface