Struct peregrine::mcus::cortexm4::core::systick::SysTick

pub struct SysTick(pub *const u8);

Pointer wrapper that implements the SysTick operations

Methods

impl SysTick

fn read_calibration(&self) -> u32

Reads the contents (as u32) of the calibration register at relative address 0x8

fn read_reload(&self) -> u32

Reads the contents (as u32) of the reload register at relative address 0x4

fn reload(&self, val: u32)

Writes the value of the given argument to the reload register at relative address 0x4 and offset 0 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn read_status_and_control(&self) -> u32

Reads the contents (as u32) of the status_and_control register at relative address 0x0

fn set_enable_bit(&self)

Writes the value 0x1 to the status_and_control register at relative address 0x0 and offset 0 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn unset_enable_bit(&self)

Writes the value 0x0 to the status_and_control register at relative address 0x0 and offset 0 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn set_exception_enable_bit(&self)

Writes the value 0x1 to the status_and_control register at relative address 0x0 and offset 1 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn unset_exception_enable_bit(&self)

Writes the value 0x0 to the status_and_control register at relative address 0x0 and offset 1 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn needs_init(&self) -> bool

states the SysTick needs to be initialized

fn init(&self)