Struct peregrine::mcus::cortexm4::kinetis::k64::K64

pub struct K64 {
    pub nvic: K64NVIC,
    pub wdog: Watchdog,
    pub systick: SysTick,
    pub fpu_coproc: Access,
    pub fpu: Unit,
    pub sim: SIM,
}

Generated definition of the K64 MCU

source: http://www.nxp.com/files/microcontrollers/doc/ref_manual/K64P144M120SF5RM.pdf

link script: MK64FN1M0VLL12.ld

Fields
nvic: K64NVIC wdog: Watchdog systick: SysTick fpu_coproc: Access fpu: Unit sim: SIM
Methods

impl K64

const fn new() -> K64

Creates a new MCU instance with all peripherals set to the correct addresses.

This is intended to be used internally, but there is nothing inherently unsafe about creating multiple instances, but thread/access safety still applies.

fn copy_data_section(&self)

Copies the .data section (specified via the data block) to main memory

fn null_bss(&self)

Zeroes all bytes in the .bss section as specified in the bss block.

fn init(&self)

Initializes the mcu given then memory addresses and peripherals provided to the macro.

The generated code is as follows:

pub fn init(&self) {
    self.copy_data_section();
    self.null_bss();
    ... for p in periphs ...
    if p.needs_init() { p.init(); }
    ...                  ...

NOTE: peripherals are initialized in the order they are defined

Each peripheral's ::needs_init() is a compile time constant, and should not be costly.