Struct peregrine::mcus::kinetis::k64::Watchdog

pub struct Watchdog(pub *const u8);

Pointer wrapper that implements the Watchdog operations

source: http://www.nxp.com/files/microcontrollers/doc/ref_manual/K64P144M120SF5RM.pdf

Methods

impl Watchdog

const DISABLED: u32 = 0u32

const ENABLED: u32 = 1u32

fn read_prescaler(&self) -> u32

Reads the contents (as u32) of the prescaler register at relative address 0x16

fn set_prescaler(&self, val: u8)

Writes the value of the given argument to the prescaler register at relative address 0x16 and offset 8 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

const REFRESH_VALUE_ONE: u32 = 42498u32

const REFRESH_VALUE_TWO: u32 = 46208u32

fn refresh(&self)

Consecutively writes the values [REFRESH_VALUE_ONE, REFRESH_VALUE_TWO] to the refresh register at relative address 0xC and offset 0 bits.

fn read_reset_count(&self) -> u32

Reads the contents (as u32) of the reset_count register at relative address 0x14

fn clear_reset_count(&self)

Writes the value 0xFFFF to the reset_count register at relative address 0x14 and offset 0 bits.

const STATUS_CONTROL_ALT_CLOCK_SRC: u32 = 1u32

const STATUS_CONTROL_BYTE_TEST: u32 = 1u32

const STATUS_CONTROL_LPO_CLOCK_SRC: u32 = 0u32

const STATUS_CONTROL_QUICK_TEST: u32 = 0u32

fn read_status_control(&self) -> u16

Reads the contents (as u16) of the status_control register at relative address 0x0

fn disable(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 0 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 0 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_alt_clock(&self)

Writes the value STATUS_CONTROL_ALT_CLOCK_SRC to the status_control register at relative address 0x0 and offset 1 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_lpo_clock(&self)

Writes the value STATUS_CONTROL_LPO_CLOCK_SRC to the status_control register at relative address 0x0 and offset 1 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn reset_breadcrumb_mode(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 2 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn reset_default_mode(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 2 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_windowing(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 3 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_windowing(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 3 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn allow_updates(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 4 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disallow_updates(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 4 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_in_debug(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 5 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_in_debug(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 5 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_in_stop(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 6 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_in_stop(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 6 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_in_wait(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 7 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_in_wait(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 7 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn run_self_test(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 10 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_byte_test(&self)

Writes the value STATUS_CONTROL_BYTE_TEST to the status_control register at relative address 0x0 and offset 11 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_quick_test(&self)

Writes the value STATUS_CONTROL_QUICK_TEST to the status_control register at relative address 0x0 and offset 11 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_byte_one_test(&self)

Writes the value 0x1 to the status_control register at relative address 0x0 and offset 12 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_byte_three_test(&self)

Writes the value 0x3 to the status_control register at relative address 0x0 and offset 12 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_byte_two_test(&self)

Writes the value 0x2 to the status_control register at relative address 0x0 and offset 12 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn use_byte_zero_test(&self)

Writes the value 0x0 to the status_control register at relative address 0x0 and offset 12 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn disable_self_test(&self)

Writes the value DISABLED to the status_control register at relative address 0x0 and offset 14 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn enable_self_test(&self)

Writes the value ENABLED to the status_control register at relative address 0x0 and offset 14 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn read_status_control_low(&self) -> u16

Reads the contents (as u16) of the status_control_low register at relative address 0x2

fn set_interrupt_flag(&self)

Writes the value ENABLED to the status_control_low register at relative address 0x2 and offset 15 bits.

NOTE: all values in unaligned writes are subject to masking which is not displayed here.

fn read_timeout(&self) -> u32

Reads the contents (as u32) of the timeout register at relative address 0x4

fn set_timeout(&self, val: u32)

Writes the value of the given argument to the timeout register at relative address 0x4 and offset 0 bits.

fn read_timer_output(&self) -> u32

Reads the contents (as u32) of the timer_output register at relative address 0x10

fn set_timer_output(&self, val: u32)

Writes the value of the given argument to the timer_output register at relative address 0x10 and offset 0 bits.

const UNLOCK_VALUE_ONE: u32 = 50464u32

const UNLOCK_VALUE_TWO: u32 = 55592u32

fn unlock(&self)

Consecutively writes the values [UNLOCK_VALUE_ONE, UNLOCK_VALUE_TWO] to the unlock register at relative address 0xE and offset 0 bits.

fn read_window(&self) -> u32

Reads the contents (as u32) of the window register at relative address 0x8

fn set_window(&self, val: u32)

Writes the value of the given argument to the window register at relative address 0x8 and offset 0 bits.