Struct peregrine::os::mman::SlabAllocator

pub struct SlabAllocator { /* fields omitted */ }

Methods

impl SlabAllocator

fn new(iov: IOVec, block_size: usize) -> SlabAllocator

Instantiates a new slab allocator managing the region defined by the IOVec.

A bitmap is used to track used blocks, thus buffer_length / block_size will not always give you the proper block count. If the buffer has enough room for the bitmap, but not another block, this dead space will be used for the bitmap.

fn alloc(&mut self, block_count: usize) -> Result<IOVec, &'static str>

Allocate the given number of blocks. The returned IOVec should be "given back" to the allocator when freed.

Errors occur from bounds checking, internal failures, or unavailability of memory.

fn free(&mut self, iov: IOVec) -> Result<(), &'static str>

Free the given region of allocated memory. The input ideally should come from a previous ::alloc().

Errors occur from bounds checking, freeing unallocated/unmanaged memory, or unaligned IOVec pointer and/or size.

NOTE: if Bitmap::checked_clear(...) returns an error, the error is immediately returned. This is being worked on but can lead to things not being freed. The solution will likely be returning the index with the error -- but we cannot return a Vec of errors for instance.

fn blocks(&self) -> usize

Get the total number of managed blocks.

fn free_blocks(&self) -> usize

Get the number of blocks available to the manager.

fn used_blocks(&self) -> usize

Get the number of free blocks in the manager.