Trait peregrine::traits::MCU

pub trait MCU {
    fn get_nvic(&self) -> &NVIC;
    fn stack_memory(&self) -> IOVec;
    fn heap_memory(&self) -> IOVec;
}

Enables generic handling and initialization of an MCU.

Required Methods

fn get_nvic(&self) -> &NVIC

Returns the NVIC interface for the MCU. The NVIC returned must satisfy the NVIC trait.

fn stack_memory(&self) -> IOVec

Gives an IOVec describing the region of memory dedicated to stack.

fn heap_memory(&self) -> IOVec

Gives an IOVec describing the region of memory dedicated to the heap.