Trait peregrine::traits::NVIC

pub trait NVIC {
    fn enable_irq(&self, irq: u8);
    fn disable_irq(&self, irq: u8);
    fn is_enabled(&self, irq: u8) -> bool;
    fn set_pending(&self, irq: u8);
    fn clear_pending(&self, irq: u8);
    fn is_pending(&self, irq: u8) -> bool;
    fn is_active(&self, irq: u8) -> bool;
    fn set_priority(&self, irq: u8, prio: u8);
    fn get_priority(&self, irq: u8) -> u8;
}

Standard interface to the Nested Vector Interrupt Controller.

Required Methods

fn enable_irq(&self, irq: u8)

Enables the given IRQ.

fn disable_irq(&self, irq: u8)

Disables the given IRQ.

fn is_enabled(&self, irq: u8) -> bool

Check whether the given IRQ is enabled.

fn set_pending(&self, irq: u8)

Set the given IRQ as pending, and to be run when priorities dictate.

fn clear_pending(&self, irq: u8)

Clear the given IRQ from the pending list.

fn is_pending(&self, irq: u8) -> bool

Check whether the given IRQ is pending.

fn is_active(&self, irq: u8) -> bool

Check if the given IRQ is currently being serviced;

fn set_priority(&self, irq: u8, prio: u8)

Set the priority of the given IRQ.

Priorities are limited to the number of priority bits, and this function should handle all shifting/masking needed.

fn get_priority(&self, irq: u8) -> u8

Gets the priority of the given IRQ.

Priorities should be returned in the range [0, priority_bits<<2].